requirements.txt
Django==2.2
django-tinymce~=2.8.0
mysqlclient~=2.2.4
numpy~=2.0.2
pandas~=2.2.3
Pillow~=9.1.1
python-dateutil~=2.9.0.post0
pytz~=2024.2
six~=1.16.0
sqlparse~=0.5.1

djangorestframework~=3.12.4
pip~=24.2
wheel~=0.43.0
filelock~=3.16.1
redis~=5.0.8
setuptools~=74.1.2
PyYAML~=6.0.2
jinja2~=3.1.4
MarkupSafe~=2.1.5
drf-yasg~=1.20.0